Create Calendar.java
Calendar class builds / refreshes the calendar as required.  It does not currently work.  The layout is functional; however, there is an NPE error occurring when trying to fill in the column headers (line 75-78).  There is a link in the initial comments that leads to a tutorial showing something similar to what we're trying to do.  I wanted to include it to ensure plagiarism isn't an issue.