Data connection code snippets
This is not a complete .java document. It contains the basic code
fragments to retrieve form data, connect to the DB and retrieve data,
assign data to variables and collect into an Array object for use. The
connection section is the primary part of the code, but use as much or
as little as you like.