See comments
Added a Logout button to the calendar.  Button is justified to the right, functional, and works correctly for now.  Only thing that needs to be added is the "disconnection from the database".

Also, tried to setup a control loop to stop the next / back buttons from going beyond 5-years, but its not catching.  I think its something to do with the specific values I'm looking for in the if-statement, but unsure.