Fixed Calendar Navigation Issue
Fixed the issue with the navigation buttons. After every time the
buttons are pressed, the code checks to see if they should be enabled
or not. This prevents us from needing an error message. It looks like
the problem with the original code was that it wasn’t stopping the
calendar from being set to the previous or next month. So, the if
conditions were only met once and the gregorianCalendar variable was
continually set to an updated value, regardless of whether it was
outside of the allowed range. On another note, I increased the allowed
range to include 5 full years passed the current year.