EventScreen
More work on EventScreen. Making the minute combo boxes always display
2 digits (e.g. 00 and 05) by storing Strings rather than Integers.
Requires conversions when going back and forth between the event and
the combo boxes. Implementing the save handler to grab the values from
the GUI and store them in an event so they can be saved in the database.